@@ -1,6 +1,6 @@
 package net.unknown.musicapi;
 
-import net.unknown.musicapi.providers.itunes.HttpClientItunesApi;
+import net.unknown.musicapi.providers.ItunesClient;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -11,7 +11,7 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 public class MusicApiApplication {
 
     @Autowired
-    private HttpClientItunesApi httpClientItunesApi;
+    private ItunesClient itunesClient;
 
     public static void main(String[] args) {
         SpringApplication.run(MusicApiApplication.class, args);
