@@ -8,28 +8,22 @@ import org.springframework.cache.caffeine.CaffeineCacheManager;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import java.net.http.HttpClient;
 import java.util.concurrent.TimeUnit;
 
 @Configuration
 @EnableCaching
-public class Beans {
+public class CacheConfiguration {
 
     @Bean
-    public HttpClient getHttpClient() {
-        return HttpClient.newHttpClient();
-    }
+    public CacheManager cacheManager(@Value("${music.provider.cache.expiration}") int expirationInMinutes) {
+        Caffeine<Object, Object> caffeine = Caffeine
+                .newBuilder()
+                .expireAfterWrite(expirationInMinutes, TimeUnit.MINUTES);
 
-    @Bean
-    public Caffeine caffeineConfig(@Value("${music.provider.cache.expiration}") int expirationInMinutes) {
-        return Caffeine.newBuilder().expireAfterWrite(expirationInMinutes, TimeUnit.MINUTES);
-    }
-
-    @Bean
-    public CacheManager cacheManager(Caffeine caffeine) {
         CaffeineCacheManager caffeineCacheManager = new CaffeineCacheManager();
         caffeineCacheManager.setCaffeine(caffeine);
         return caffeineCacheManager;
     }
+
 
 }
