@@ -1,7 +1,6 @@
 package net.unknown.musicapi.configuration;
 
-import net.unknown.musicapi.security.UserIdFilter;
-import org.springframework.beans.factory.annotation.Autowired;
+import net.unknown.musicapi.filters.UserIdFilter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
@@ -10,11 +9,9 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 public class SecurityConfiguration {
 
-    @Autowired
-    private UserIdFilter userIdFilter;
-
-    @Bean
-    public FilterRegistrationBean<UserIdFilter> securityFilter( @Value("${security.url.pattern}") String urlPattern) {
+    @Bean("secure")
+    public FilterRegistrationBean<UserIdFilter> securityFilter(
+            @Value("${security.url.pattern}") String urlPattern, UserIdFilter userIdFilter) {
         FilterRegistrationBean<UserIdFilter> registrationBean = new FilterRegistrationBean<>();
         registrationBean.setFilter(userIdFilter);
         registrationBean.addUrlPatterns(urlPattern);
