@@ -1,7 +1,7 @@
 package net.unknown.musicapi.persistence.repositories;
 
-import net.unknown.musicapi.persistence.models.User;
+import net.unknown.musicapi.persistence.models.Fan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepo extends JpaRepository<User, Long> {
+public interface UserRepo extends JpaRepository<Fan, Long> {
 }
