@@ -1,20 +1,17 @@
 package net.unknown.musicapi.services;
 
-import net.unknown.musicapi.controllers.dtos.ArtistDto;
+import net.unknown.musicapi.dtos.ArtistDto;
 import net.unknown.musicapi.persistence.models.Artist;
-import net.unknown.musicapi.persistence.models.User;
 import net.unknown.musicapi.persistence.repositories.ArtistRepo;
 import net.unknown.musicapi.persistence.repositories.UserRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.Optional;
-
 @Service
 public class UserService {
 
-    private UserRepo userRepo;
-    private ArtistRepo artistRepo;
+    private final UserRepo userRepo;
+    private final ArtistRepo artistRepo;
 
     @Autowired
     public UserService(UserRepo userRepo, ArtistRepo artistRepo) {
@@ -23,22 +20,19 @@ public class UserService {
     }
 
     public void saveArtistForUser(ArtistDto artistDto, long userId) {
-        Optional<User> wrappedUser = userRepo.findById(userId);
-
-        if (wrappedUser.isPresent()) {
-            User user = wrappedUser.get();
-            Artist artist = new Artist(artistDto.getArtistId(), artistDto.getArtistName(), artistDto.getAmgArtistId());
-            artist.addUser(user);
-            artistRepo.save(artist);
-        }
+        userRepo.findById(userId)
+                .ifPresent(user -> {
+                    Artist artist = new Artist(artistDto);
+                    artist.addUser(user);
+                    artistRepo.save(artist);
+                    user.addArtist(artist);
+                    userRepo.save(user);
+                });
     }
 
     public boolean isFavoriteArtist(long amgArtistId, long userId) {
-        Optional<User> wrappedUser = userRepo.findById(userId);
-        return wrappedUser.map(user -> user.isArtistPresent(amgArtistId)).orElse(false);
-    }
-
-    public boolean isNotFavoriteArtist(long amgArtistId, long userId) {
-        return !isFavoriteArtist(amgArtistId, userId);
+        return userRepo
+                .findById(userId)
+                .map(user -> user.isArtistPresent(amgArtistId)).orElse(false);
     }
 }
