@@ -1,5 +1,6 @@
-package net.unknown.musicapi.providers.itunes;
+package net.unknown.musicapi.providers;
 
+import net.unknown.musicapi.configuration.ItunesApiConfiguration;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,16 +14,16 @@ import org.springframework.test.context.ActiveProfiles;
 public class ItunesApiUrisTest {
 
     @Autowired
-    private ItunesApiUris itunesApiUris;
+    private ItunesApiConfiguration itunesApiUris;
 
     @Test
-    public void should_returnCorrectFetchUrl_whenGiveCorrectParams() {
+    public void constructUriForTop5AlbumsRetrieval() {
         String artistId = "123";
         Assertions.assertThat(itunesApiUris.getTopArtistAlbums(artistId)).isEqualTo("https://localhost.com/lookup?entity=album&limit=5&amgArtistId=" + artistId);
     }
 
     @Test
-    public void should_returnCorrectSearchUrl_whenGiveCorrectParams() {
+    public void constructUriForSearchingArtistByKeyword() {
         String artistId = "123";
         Assertions.assertThat(itunesApiUris.getUriToSearchArtist(artistId)).isEqualTo("https://localhost.com/search?entity=allArtist&term=" + artistId);
     }
